Character/Enemy classes
Put in the classes I described at the meeting. Just have the
playercontroller script inherit from Character and it should be able to
move without the move method. But you will need to call tilemap.FindPath
before calling the move method. That's so I can modify the path when
need be in the enemy code.