Fixing enemy turn changing
in player manager: the enemy/player updates are called in manager's
lateupdate to ensure a target was picked before the enemy tries to move.
Enemy now finds target not end and resets when ending enemy turn rather
than setting active to true.

In enemy class: now finds path to target and sets end to 1 minus that
(not the other way around. Attacks instead of moving if range+1 tiles in
the tilelist. no longer attacks based on vector3 distance.