Animations and Weapons
Added in animations for the player. Include an idle, a walk, and a weapon hold animation. The character has a weapon position empty game object. When equipping a weapon just make it the parent and have a position of 0,0,0. This should make the weapon follow the animation.

All weapon types have prefabs. Upgraded weapons just have to have the sprite changed on the prefab of the same type. For example, change the sword sprite for an upgraded sword.