A working build with external files
Changed the paths for XML files in all scripts to now pull from a
Resources/XML/ Folder, as Unity creates a resources folder even during
builds. This allows for a working build that sees all scripts. Also
altered some values in there to provide more of a starting off feeling.